fix(core): validate priority range in phase 9.5 after capability guards
Gate: G3 Candidate
Slice: cross-layer
Impact: core-api=changed; data-format=none; adapter-api=none
Scope: 与 regulation/timeWindows 同一「校验时机」原则收口：wire priority 改 Option<i64> 保留原始整数（JSON integer type 检查留在 wire，小数仍 JsonShape），i32 语义范围移入 AccessRegistry::try_new phase 9.5（capability guard 之后）；AccessRule priority 存储/with_priority/priority() 改 i64；phase 9.5 合并为单一 input-order 循环（逐规则先 priority 后 regulation，严格遵守同 phase 首错顺序）；新增 CoreError::InvalidAccessRulePriority{priority}，loader 按值归因 accessRules[i].priority；core/loader 各补 guard-优先、范围拒绝、phase 内顺序与边界回归测试；SSOT phase 9 文档同步
Validation: cargo +1.96.0 fmt --all -- --check; cargo +1.96.0 check --workspace --all-targets（零警告）; cargo +1.96.0 test --workspace --locked（全绿）
Docs: updated

Refs: #262

+++ crates/laneflow-data/src/wire.rs
@@ -200,7 +200,10 @@ pub(crate) struct WireAccessRule {
     #[serde(default, deserialize_with = "non_null_option")]
     pub(crate) regulation: Option<WireRegulation>,
     #[serde(default, deserialize_with = "non_null_option")]
-    pub(crate) priority: Option<i32>,
+    // 与 timeWindows 分钟字段同理：JSON integer type 检查留在 wire（小数仍
+    // JsonShape 拒绝），i32 语义范围保留原始整数值，由 Core phase 9.5 在
+    // capability guard 之后校验。
+    pub(crate) priority: Option<i64>,
 }
 
 #[derive(Deserialize)]
